+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix logic how locators are determined for finding groups.
+
 ## [0.5.2] - 2025-12-01
 
 ### Fixed